+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Fixed
  - Various compiler error reporting improvements (#1810)
+ - Fixed cache leak in agent when deployments are canceled (#1883)
  - Improved robustness of modules update (#1885)
 
 ## Added